part-2/requirements.txt
google-adk
serpapi